Included buffer to split the item set
The buffer operator is included (line 7) which splits the source channel into subsets. Here the set of items is split in two. The buffer is included to set the number of the parallel treads by hand, so that the speed can be compared between using different numbers of parallel processes.