Create the calculatePlog process
The main aim of this commit is to update the process where the wikidata and smiles set are printed into a setup for a process which calculates the Plog of the molecules belonging to the SMILES strings (in set 'smiles'). Therefore the following changes are made:
line 10: change the name of the process to calculatePlog, to make fit with the purpose of the process.
line 12: make the input more general and take each set from the molecules_ch source.
line 15: print a text to make the user aware when the process is still being executed.
line 16-21: for-loop which takes the smiles set and calls for each entry in the set the molecule CDK object and stores them in 'mol' (line 20).