Conversion CDK molecule object to IAtomContainer
The main part of this commit is the addition of the conversion of the CDK molecule object to the IAtomContainer object. This is done to make to make it compatible with the function to calculate the Plog of the molecules in the set. Additionally the documentation of the parts in the process is improved, to make it easier for the user to understand the steps of the process.