Addition logP calculation
The main part of this commit is the addition of the calculation of the logP values of the IAtomContainer objects (Iatom). The JPlogPDescriptor from the CDK library is used herefore, because it takes an IAtomContainer object and outputs the logP value of the molecule. Other updates in this commit are the addition of the loading of the libraries at the beginning of the file, to make the process executable. The 'output' command within the process is changed to 'exec' to make that part of the file execute.