Addition of loop over number of CPUs
The focus of this commit on making a loop that runs the process in parallel for different numbers of CPUs. The buffersize is made dependent on the number of CPUs that are used (i.e. the buffer size changes every iteration). The buffersize determines the number of CPUs the process is being split over. To measure the time duration of one iteration of the loop the current time at the start and end of the loop is taken and the difference is calculated at the end and printed.