Print CPU number and duration to file
In this commit the time duration of each iteration (each iteration the number of CPUs that the process is run on changes) are stored in a .tsv file called CPU_duration.tsv. This is done to provide the user with an easy to interpret file that can be used for further processing. The 'for-loop' is changed in an '.upto()' command which loops through the process for the number of CPUs (here 12) iterations. This is changed because storing of the right indexing is more robust than if a for-loop was used.