Add getSMILES.rq
The file getSMILES.rq is included in the repository. This file contains the wikidata query that can be used to request the full list of molecules and SMILES from the wikidata database. The origin of this file is in the repository 'scientificProgramming' of Egon Willighagen which is a dedicated repository for the course this assignment is a part of.