Add MoleculesSmiles.tsv
This file stores the query results from the wikidata query that is stored in 'getSMILES.rq'. The code that is within this file is copied to the Wikidata query service (https://query.wikidata.org/) and the response is saved in MoleculesSmiles.tsv which is here uploaded to the repository. The response contains 3 columns named compound, smiles and isoSmiles respectively. And they contain the wikidata entry url of the molecule, the canonical SMILES string and the isomeric SMILES string respectively.