Introduction, Installation, Requirements
In this commit an introduction is written for this repository. Explaining what the goal is of this project to make the user aware of the steps that are taken to generate results and what results are to be obtained.
Furthermore, the installation requirements are specified, to make the user able to install the project on its own computer and to run it with user specific settings.