Execution, Introduction and Results
The Execution part was included. Here it is explained what the user should do to run the experiment on her/his own device. Furthermore the introduction saw minor changes and the results section was extended with an explanation of where the results of the experiment can be found.
Furthermore, the authors section was updated as part of the repository is not just written by the main author.