Run parallel process on large dataset
In this commit the large dataset 'MoleculesSmiles.tsv' is used in the calculation of logP values. This file holds both canonical and isomeric SMILES strings. Within the process calculatelogP both canonical and isomeric SMILES are input. First the process tries to find the CDK molecule object of the isomeric SMILES, if it does not succeed it tries to find the CDK molecul object for canonical SMILES. If this is not successful a warning message is printed to indicate that for that entry in the set there is no logP value calculated.