Duration of process on different CPU settings
In this commit the file 'CPU_duration.tsv' is added to the repository. This file contains the 2 columns with the number of CPUs that are used for an iteration of the calculation and the duration of that process respectively. This file will be used in an R markdown file to present these results.