Add results files
In this commit the R markdown file 'Assignment3.Rmd' and 'index.html' are added. The first file processes the 'CPU_duration.tsv' file in R markdown format and includes an introduction and discussion of the experiment. 'index.html' is included so a Github Page can be made and the results can be presented in a modern internet browser without needing to install other dependencies, this is convenient for the user. These files are included because it gives a clear overview of what is done and includes a discussion of what should happen to the project to improve it beyond its current state.